+++ src/db/user_db.c
@@ -884,7 +884,7 @@ int user_db_get_all_format(database *db, char *buffer, size_t buffer_size) {
     }
 
     sqlite3_finalize(stmt);
-    return written;
+    return (int)written;
 }
 
 char *user_db_get_all_format_old(database *db) {
