+++ src/ui/screens/ui_create_user.c
@@ -614,7 +614,7 @@ static void handle_get_all_button(struct ui_create_user *ui, database *user_db) 
     }
 
     // 512 for header + 512 for each row
-    size_t buffer_size = 512 + 512 * total_users;
+    size_t buffer_size = 512 + 512 * (size_t)total_users;
 
     ui->str_table_content = malloc(buffer_size);
     if (!ui->str_table_content) {
@@ -632,8 +632,8 @@ static void handle_get_all_button(struct ui_create_user *ui, database *user_db) 
     // Set the panel_content_bounds rectangle based on the width and height of the retrieved text
     if (ui->str_table_content) {
         Vector2 text_size = MeasureTextEx(GuiGetFont(), ui->str_table_content, FONT_SIZE, 0);
-        ui->sp_table_view.panel_content_bounds.width = text_size.x * 0.9;
-        ui->sp_table_view.panel_content_bounds.height = text_size.y / 0.7;
+        ui->sp_table_view.panel_content_bounds.width = text_size.x * 0.9F;
+        ui->sp_table_view.panel_content_bounds.height = text_size.y / 0.7F;
     }
 
     user_db_get_all(user_db);
